Move MenuFeatureSelectionController to landing.features package and rename.
