+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/landing/features/LandingMenuFeatureSelectionController.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot.feature_selection;
+package com.novoda.tpbot.landing.features;
 
 import android.content.Context;
 import android.support.annotation.MenuRes;
@@ -9,8 +9,11 @@ import android.view.MenuItem;
 
 import com.novoda.tpbot.R;
 import com.novoda.tpbot.bot.video.calling.VideoCallSharedPreferencesPersistence;
+import com.novoda.tpbot.feature_selection.FeatureSelectionController;
+import com.novoda.tpbot.feature_selection.FeatureSelectionPersistence;
+import com.novoda.tpbot.feature_selection.ServerConnectionSharedPreferencesPersistence;
 
-public final class MenuFeatureSelectionController implements FeatureSelectionController<Menu, MenuItem> {
+public final class LandingMenuFeatureSelectionController implements FeatureSelectionController<Menu, MenuItem> {
 
     @MenuRes
     private static final int FEATURE_MENU_RESOURCE = R.menu.feature_menu;
@@ -25,10 +28,10 @@ public final class MenuFeatureSelectionController implements FeatureSelectionCon
         features.put(R.id.video_call_menu_item, VideoCallSharedPreferencesPersistence.newInstance(context));
         features.put(R.id.server_connection_menu_item, ServerConnectionSharedPreferencesPersistence.newInstance(context));
 
-        return new MenuFeatureSelectionController(menuInflater, features);
+        return new LandingMenuFeatureSelectionController(menuInflater, features);
     }
 
-    private MenuFeatureSelectionController(MenuInflater menuInflater, SparseArray<FeatureSelectionPersistence> features) {
+    private LandingMenuFeatureSelectionController(MenuInflater menuInflater, SparseArray<FeatureSelectionPersistence> features) {
         this.menuInflater = menuInflater;
         this.features = features;
     }
